requirements.txt
streamlit
openai
python-dotenv
tenacity
requests